\name{binary.comp.methods}
\alias{binary.comp.methods}
\title{
Binary comparison methods
}
\description{
This function provides the methods implemented for calculating binary similarity between hotspots obtained from different sampling schemes and those obtained from the baseline (reference) sampling scheme. It is used by functions \code{\link{binary.comparison}}, \code{\link{sequential.corr}} and \code{\link{hotspots.comparison}}.
}
\usage{
binary.comp.methods()
}
\details{
}
\value{
A character vector naming the methods implemented.
}
\references{
}
\author{
A. Marcia Barbosa
}
\note{
}
\seealso{
\code{\link{binary.comparison}}, \code{\link{sequential.corr}}
}
\examples{
binary.comp.methods()
}
