\name{DeadCanMove-package}
\alias{DeadCanMove-package}
\alias{DeadCanMove}
\docType{package}
\title{
Assess how spatial roadkill patterns change with temporal sampling scheme
}
\description{
From a baseline data frame of dead individuals recorded daily at different 
road stretches, simulate varying sub-sampling schemes, calculate roadkill 
hotspots based on each scheme, and compare hotspot patterns between sampling 
schemes and the baseline data.
}
\details{
\tabular{ll}{
Package: \tab DeadCanMove\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-04-17\cr
License: \tab GPL-3\cr
}
}
\author{
Barbosa A.M., Marques J.T., Santos S.M., Lourenco A., Medinas D., Beja P., 
Mira A.

Maintainer: A. Marcia Barbosa <barbosa@uevora.pt>
}
\references{
Santos S.M., Marques J.T., Lourenco A., Medinas D., Barbosa A.M., Beja P., 
Mira A. (under revision) Dead can move? Effects of sampling frequency on the 
identification of roadkill hotspots
}
\keyword{ package }
\seealso{
}
\examples{
}
